Added blast propagation to all links
