+++ sources/clusterrole_test.go
@@ -2,47 +2,29 @@ package sources
 
 import (
 	"testing"
-
-	"github.com/overmindtech/sdp-go"
 )
 
-var clusterRoleBindingYAML = `
+var clusterRoleYAML = `
 apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
+kind: ClusterRole
 metadata:
-  name: admin-binding
-subjects:
-- kind: Group
-  name: system:serviceaccounts:default
-  apiGroup: rbac.authorization.k8s.io
-roleRef:
-  kind: ClusterRole
-  name: admin
-  apiGroup: rbac.authorization.k8s.io
+  name: read-only
+rules:
+- apiGroups: [""]
+  resources: ["*"]
+  verbs: ["get", "list", "watch"]
+
 `
 
-func TestClusterRoleBindingSource(t *testing.T) {
-	source := newClusterRoleBindingSource(CurrentCluster.ClientSet, CurrentCluster.Name, []string{})
+func TestClusterRoleSource(t *testing.T) {
+	source := newClusterRoleSource(CurrentCluster.ClientSet, CurrentCluster.Name, []string{})
 
 	st := SourceTests{
-		Source:    source,
-		GetQuery:  "admin-binding",
-		GetScope:  CurrentCluster.Name,
-		SetupYAML: clusterRoleBindingYAML,
-		GetQueryTests: QueryTests{
-			{
-				ExpectedType:   "ClusterRole",
-				ExpectedMethod: sdp.QueryMethod_GET,
-				ExpectedQuery:  "admin",
-				ExpectedScope:  CurrentCluster.Name,
-			},
-			{
-				ExpectedType:   "Group",
-				ExpectedMethod: sdp.QueryMethod_GET,
-				ExpectedQuery:  "system:serviceaccounts:default",
-				ExpectedScope:  CurrentCluster.Name,
-			},
-		},
+		Source:        source,
+		GetQuery:      "read-only",
+		GetScope:      CurrentCluster.Name,
+		SetupYAML:     clusterRoleYAML,
+		GetQueryTests: QueryTests{},
 	}
 
 	st.Execute(t)
