+++ sources/endpoints.go
@@ -26,6 +26,11 @@ func EndpointsExtractor(resource *v1.Endpoints, scope string) ([]*sdp.LinkedItem
 						Query:  address.Hostname,
 						Type:   "dns",
 					},
+					BlastPropagation: &sdp.BlastPropagation{
+						// Always propagate over DNS
+						In:  true,
+						Out: true,
+					},
 				})
 			}
 
@@ -37,6 +42,12 @@ func EndpointsExtractor(resource *v1.Endpoints, scope string) ([]*sdp.LinkedItem
 						Method: sdp.QueryMethod_GET,
 						Query:  *address.NodeName,
 					},
+					BlastPropagation: &sdp.BlastPropagation{
+						// Changes to the node can affect the endpoint
+						In: true,
+						// Changes to the endpoint cannot affect the node
+						Out: false,
+					},
 				})
 			}
 
@@ -47,6 +58,11 @@ func EndpointsExtractor(resource *v1.Endpoints, scope string) ([]*sdp.LinkedItem
 						Method: sdp.QueryMethod_GET,
 						Query:  address.IP,
 						Scope:  "global",
+					},
+					BlastPropagation: &sdp.BlastPropagation{
+						// Always propagate over IP
+						In:  true,
+						Out: true,
 					},
 				})
 			}
