+++ sources/service.go
@@ -21,6 +21,12 @@ func serviceExtractor(resource *v1.Service, scope string) ([]*sdp.LinkedItemQuer
 				}),
 				Scope: scope,
 			},
+			BlastPropagation: &sdp.BlastPropagation{
+				// Bidirectional propagation since we control the pods, and the
+				// pods host the service
+				In:  true,
+				Out: true,
+			},
 		})
 	}
 
@@ -44,6 +50,11 @@ func serviceExtractor(resource *v1.Service, scope string) ([]*sdp.LinkedItemQuer
 					Query:  ip,
 					Scope:  "global",
 				},
+				BlastPropagation: &sdp.BlastPropagation{
+					// IPs are always bidirectional
+					In:  true,
+					Out: true,
+				},
 			})
 		}
 	}
@@ -56,6 +67,11 @@ func serviceExtractor(resource *v1.Service, scope string) ([]*sdp.LinkedItemQuer
 				Query:  resource.Spec.ExternalName,
 				Scope:  "global",
 			},
+			BlastPropagation: &sdp.BlastPropagation{
+				// DNS is always bidirectional
+				In:  true,
+				Out: true,
+			},
 		})
 	}
 
@@ -66,6 +82,12 @@ func serviceExtractor(resource *v1.Service, scope string) ([]*sdp.LinkedItemQuer
 			Method: sdp.QueryMethod_GET,
 			Query:  resource.Name,
 			Scope:  scope,
+		},
+		BlastPropagation: &sdp.BlastPropagation{
+			// The service causes the endpoint to be created, so changes to the
+			// service can affect the endpoint and vice versa
+			In:  true,
+			Out: true,
 		},
 	})
 
@@ -78,6 +100,11 @@ func serviceExtractor(resource *v1.Service, scope string) ([]*sdp.LinkedItemQuer
 					Query:  ingress.IP,
 					Scope:  "global",
 				},
+				BlastPropagation: &sdp.BlastPropagation{
+					// IPs are always bidirectional
+					In:  true,
+					Out: true,
+				},
 			})
 		}
 
@@ -88,6 +115,11 @@ func serviceExtractor(resource *v1.Service, scope string) ([]*sdp.LinkedItemQuer
 					Method: sdp.QueryMethod_GET,
 					Query:  ingress.Hostname,
 					Scope:  "global",
+				},
+				BlastPropagation: &sdp.BlastPropagation{
+					// DNS is always bidirectional
+					In:  true,
+					Out: true,
 				},
 			})
 		}
